requirements.txt
contourpy==1.3.2
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyclipper==1.3.0.post6
pyparsing==3.2.3
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.2
setuptools==80.3.1
shapely==2.1.0
six==1.17.0
sympy==1.14.0
tifffile==2025.3.30
torch==2.7.0
torchvision==0.22.0
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
SQLAlchemy==2.0.25
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
